feat(web): user can choose what game he/she wants to play
